Implement full dynamic matchcard creation
This commit completes the dynamic matchcard creation feature-- each card (with the exception of an "other" card created for containing David's popup functionality) is created on the fly by parsing through the JSON object which holds all match information. Other changes include defaulting the border to be blue instead of green (panel-info instead of panel-success).