Finalize Backend API
Update the AppController to properly return responses in JSON format, handle a few small error cases for exceptions, and clean up a couple things. Expose the ID on the Persistable class so that GSON sends the ID to the front-end. Finally, set up the database to load the ladder when starting the server, so IDs are populated.