Setup SSL/HTTPS
This will give security errors when loaded in Chrome, since the certificate is not signed by a Certificate Authority. Once we have a domain name, we can get a cert from Let's Encrypt that will not give these errors.

Also includes small refactor of response returned from login endpoints to format it in JSON.