Fix changing status exception, and new pair existingId logic error
Changing a pair's status would always fail because the code would catch an exception related to the newPosition parameter being null (which is irrelevant when changing status), so the code now handles the cases independently, allowing the status to be changed without causing an exception in the newPosition.

Adding a new pair resulted in failure because the existingId (which was not passed) would default to 0, and thus be "valid" and cause a replacement with the existing player (who didn't really exist). The API will now pass a -1 in the existingId field if it's a new player, and the AppController will ignore existingIds of less than 0.