Bind variable CSS classes to data object
For the penalty buttons and match panels, the css classes applied depend
on the data of the object (eg. match panel is green if isDone). Use
Vue.js's v-bind to bind a class object to the class attribute of the
html element, so when data changes we can instantly update the css
classes.