Add validation for results input
Results are now input into a 2D array belonging to the match object.
This object is watched, and every edit triggers a check. If the check
fails, a warning message is displayed on the modal, and the Save Changes
button is disabled. Once the check passes, a success message is
displayed and the Save Changes button is enabled.