Add GameSession parameter to all DBManager calls and API requests
By injecting the GameSession dependency on each DBManager call that manipulates a GameSession, the DBManager is robust enough to handle manipulation of past GameSessions as well as the latest GameSession. The API now requires a GameSession to be specified (the keywords "latest" and "previous" are the only two options as of this change), opening up the front-end to the possibility of fetching and manipulating a previous GameSession.