Fix adding pair
Sanitizing the phone number on the front end was resulting in an error, so the API call was never made. However, once that was fixed, the addNewPairAtEnd function was not working, since it tried adding a new pair at an index that didn't exist. This change uses the Ladder's insertAtEnd function to properly insert the pair to the end of the ladder.