Add tab component to keep track of active tab
Tab component now keeps track of which match tab is open, so we can call
the reorderLadder API call on the correct GameSession. Also refactored
the Matches component to take in an element, so we instantiate 2 Matches
components (one for previous week, one for current week) seperately
instead of making essentially 2 copies of the component inside one
object.