Fix fragment instance, complete refactor of Matches component
Refactor the Matches component to be one large component consisting of
the tabs on the top level, containing the two Matches components as
children. This allows for events to be passed up/down the hierarchy to
change modes and pass data, which is a lot more elegant than the
previous method of changing data and watching other components data for
changes. Also moved all component related functions inline with their
declaration in the component's JSON for readability and abstraction, and
the Matches prototype functions are all now special methods to be used
by stuff outside of the Match component itself. Basically, the
component's methods should only modify itself, and the Match class
functions are there to hook in to events that will trigger something in
the component.