Fix match cards refreshing after results input
The match cards were not properly being refreshed when the match results
were being input. The fix was to pass in the gameSession when refreshing
the matches. This also revealed the reorderLadderButton was not being
updated correctly, due to the id of the HTML element missing, which has
now been added.