Only show playingstatus for current user
On the backend, when logging in the SessionResponse now contains the
id for the Player that's associated with the User that logged in. The
front end stores this as a cookie, and uses it to create a boolean on
every pair, which is then used in ladder.html to show or not show the
playing status.