Create Composer to write message for notifications, refactor classes into new package
First, classes have been moved into their own package (notifications).

Second, the Composer interface was created to allow for the possibility of different types of messages in the future (HTML formatted emails, for example), and to abstract the string operations and template definitions away. SimpleComposer composes a simple subject and message, containing things such as the current date, the player's name and the time they are scheduled to play (according to their scorecard).

Finally, the Time enum, which the Scorecard holds, has been modified to hold a string detailing the human-readable time corresponding to each time slot.