Change how database transactions are made
Create TransactionManager to handles the creation and closing of sessions and transactions. This functionality is exposed by the executeTransaction method, which can be passed a lambda and allows use of a session and transaction variable, and allows any object to be returned.

All DBManager functions have been changed to use this new method.