Prepopulate the scorecards rankings in edit modal
The scorecard rankings in the edit modal should be prepopulated to make
it obvious to the user what the previously entered ranking was. You can
implicitly know what it was (since the pairs are in the order of the
ranking that was last saved), but this change explicitly shows the last
saved ranking by setting the values of the results array (created in
api.js) initially instead of initializing them to 0.