Fix position validity check
If the user wants to place a pair at a specific position in the ladder, the valid positions range from 1 to the number of pairs in the ladder, since we will then subtract 1 to get the index with which we will do the actual ladder insertion. Since this check allowed numbers equal to 0, we would have added the pair at index -1. The front-end currently blocks the ability to specify 0, but the back-end should still be fixed to not allow it.