Implement stale session check and automatic refresh
The open/close session workflow was effective, but slowed down every transaction since it required making a new session.

This change uses a session until it notices it's stale (ie the session throws an exception when trying to open a transaction). When it's stale, we refresh the session and try again. This reduces redundant session refreshing/creation, so the only delay on transactions occurs when a session needs to be refreshed.