Fix value check
Instead of explicitly checking for null or empty string, we can just let JS test if the value is falsy (which both of those things are). Also, the name and email assignments were failing if editPlayer was null (since it relies on that object for the data), so it has been moved into the if statement.