Fix indeterminate behaviour
session.load does not handle the case where an object may not exist. It loads an object no matter what, and that object contains null data if the object doesn't exist in the database. This was causing weird test behaviour that did not appear to be correct. By changing to the session.get call, we will load the object if it exists, or get null otherwise.