Add announcement header and fix bug
Added header text to the announcement page to make it obvious for the user what the page is (since there is no indication otherwise). Also, fixed a bug where the database would get a null id after trying to insert an announcement that had too much text for the column. This was fixed by specifying the "text" type for the announcement title/message, which the database supports for storing long strings.