Initialize the NotificationManager when the jar loads
The jar that runs on the server loads from JarEntry.java, not Main.java. Since the NotificationManager was only being started inside Main.java (which is also what is used for all testing), and didn't have this initialization in JarEntry.java, the server app never even properly initialized a NotificationManager to send emails (whoops!). This should fix the emails not sending issue.