Fix Edit Player tab issues
The Edit Player tab was not working properly, because it was previously using the Vue component of the Register User tab. This change creates a new Vue component for the Edit Player tab to use, and adds a vue validator to the form (since previously it was also sharing validators).